requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
botcity-framework-base==0.4.4
botcity-framework-web==0.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.6
et-xmlfile==1.1.0
func_timeout==4.3.5
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
msedge-selenium-tools==3.141.3
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
psutil==6.0.0
pynput==1.7.7
pyperclip==1.9.0
python-dotenv==1.0.1
python-xlib==0.33
pytz==2024.1
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.2
requests==2.32.3
selenium==3.141.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==1.26.19
webdriver-manager==4.0.2
